requirements.txt
allure-pytest==2.11.1
allure-python-commons==2.11.1
pip==22.3.1
pytest==7.2.0
selenium==4.6.0
webdriver-manager==3.8.4
pytest-xdist==3.0.2

